Move hostname field tests to a test folder
