+++ frontend/src/metabase/databases/components/DatabaseHostnameWithProviderField/test/provider-config.mock.ts
